+++ src/main/scala/scodec/codecs/package.scala
@@ -536,7 +536,35 @@ package object codecs {
    * @param codec codec to encode/decode a single element of the sequence
    * @group combinators
    */
-  def repeated[A](codec: Codec[A]): Codec[collection.immutable.IndexedSeq[A]] = new IndexedSeqCodec(codec)
+  @deprecated("Use vector codec or list codec instead.", "1.3.0")
+  def repeated[A](codec: Codec[A]): Codec[collection.immutable.IndexedSeq[A]] =
+    new VectorCodec(codec).xmap[collection.immutable.IndexedSeq[A]](identity, _.toVector)
+
+  /**
+   * Codec that encodes/decodes a `Vector[A]` from a `Codec[A]`.
+   *
+   * When encoding, each `A` in the vector is encoded and all of the resulting vectors are concatenated.
+   *
+   * When decoding, `codec.decode` is called repeatedly until there are no more remaining bits and the value result
+   * of each `decode` is returned in the vector.
+   *
+   * @param codec codec to encode/decode a single element of the sequence
+   * @group combinators
+   */
+  def vector[A](codec: Codec[A]): Codec[Vector[A]] = new VectorCodec(codec)
+
+  /**
+   * Codec that encodes/decodes a `List[A]` from a `Codec[A]`.
+   *
+   * When encoding, each `A` in the list is encoded and all of the resulting vectors are concatenated.
+   *
+   * When decoding, `codec.decode` is called repeatedly until there are no more remaining bits and the value result
+   * of each `decode` is returned in the list.
+   *
+   * @param codec codec to encode/decode a single element of the sequence
+   * @group combinators
+   */
+  def list[A](codec: Codec[A]): Codec[List[A]] = new ListCodec(codec)
 
   /**
    * Combinator that chooses amongst two codecs based on an implicitly available byte ordering.